third_party/WebKit/LayoutTests/external/wpt/css/css-transforms/2d-rotate-js-expected.txt

This is a testharness.js-based test.
FAIL JS test: Rotate via javascript must show the correct computed rotation assert_equals: expected "rotate(30deg)" but got "matrix(0.866025, 0.5, -0.5, 0.866025, 0, 0)"
Harness: the test ran to completion.
